Change the Checkbox remote interrupt menu text to something more clear (#540)
* Change the menu text to something more clear

Minor: Fix a small edge case in test kill

* Changed execution to error out always on missing pid

This is to prevent Checkbox from crashing when selecting the "Stop the test case in progress and move on to the next" option.

+++ checkbox-ng/plainbox/impl/execution.py
@@ -373,6 +373,11 @@ class UnifiedRunner(IJobRunner):
                             "{}.record.gz".format(slugify(job.id)))
 
     def send_signal(self, signal, target_user):
+        if not self._running_jobs_pid:
+            # this can happen because the kill command is issued
+            # just as the job finishes
+            logger.error("No job is currently running")
+            return
         if not target_user:
             os.kill(self._running_jobs_pid, signal)
         else:
